Added two new properties to the MachineCode Class: 1. isLabel Flag to indicate whether the instruction is just a label. 2. isJump Flag to indicate whether or not the instruction is a type of Jump.
Added self. to the translate call in the IfStmt class in the translate method.
Added the necessary flags when adding jumps and labels to the instructions in both WhileStmt and IfStmt
Modified IfStmt,WhileStmt and StmtList to return the last storage location along with instructions.

MOdified the translate method in Program to accept the tuple return type from the StmtList.
MOdified the link method in Program to not attempt to link an address to any Jump or Labels.